Fix curl resource cleanup

+++ sdmx-dl-provider-ri/src/main/java/internal/util/http/curl/CurlHttpURLConnection.java
@@ -58,7 +58,7 @@ final class CurlHttpURLConnection extends HttpURLConnection {
 
     @Override
     public void disconnect() {
-        if (body != null) {
+        if (body != null && Files.exists(body)) {
             try {
                 Files.delete(body);
             } catch (IOException ex) {
